Update README doc, new task and fixes  - Adding remote console  - Fixed error with bundler recipes

+++ recipes/bundler.rb
@@ -2,19 +2,22 @@ namespace :bundler do
   task :create_symlink, :roles => :app do
     shared_dir = File.join(shared_path, 'bundle')
     release_dir = File.join(current_release, '.bundle')
-    run("mkdir -p #{shared_dir} && ln -s #{shared_dir} #{release_dir}")
+    
+    unless File.exists?(release_dir)
+      run("mkdir -p #{shared_dir} && ln -s #{shared_dir} #{release_dir}")
+    end
   end
   
   task :bundle_new_release, :roles => :app do
-    bundler.create_symlink
-    run "cd #{release_path} && bundle install --without test"
+    # bundler.create_symlink
+    run "cd #{current_path} && bundle install"
   end
   
   task :lock, :roles => :app do
-    run "cd #{current_release} && bundle lock;"
+    run "cd #{current_path} && bundle lock;"
   end
   
   task :unlock, :roles => :app do
-    run "cd #{current_release} && bundle unlock;"
+    run "cd #{current_path} && bundle unlock;"
   end
 end
